+++ projects/ids-enterprise-ng/src/lib/datagrid/soho-datagrid.d.ts
@@ -301,7 +301,6 @@ interface SohoDataGridModifiedRow {
   cells: SohoDataGridDirtyCell[];
 }
 
-
 interface SohoDataGridDirtyCell {
   value: any;
   coercedVal: any;