+++ projects/ids-enterprise-ng/src/lib/homepage/soho-homepage.component.ts
@@ -20,9 +20,6 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
 
   @Input() set homePageOptions(homePageOptions: SohoHomePageOptions) {
     this._homePageOptions = homePageOptions;
-    if (this.jQueryElement) {
-      // this.homepage.updated(); ???
-    }
   }
   get homePageOptions(): SohoHomePageOptions {
     return this._homePageOptions;
@@ -33,8 +30,8 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
    */
   @Input() set columns(columns: number) {
     this._homePageOptions.columns = columns;
-    if (this.homePage) {
-      this.homePage.settings.columns = columns;
+    if (this.homepage) {
+      this.homepage.settings.columns = columns;
     }
   }
 
@@ -47,8 +44,8 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
    */
   @Input() set gutterSize(gutterSize: number) {
     this._homePageOptions.gutterSize = gutterSize;
-    if (this.homePage) {
-      this.homePage.settings.gutterSize = gutterSize;
+    if (this.homepage) {
+      this.homepage.settings.gutterSize = gutterSize;
     }
   }
 
@@ -61,8 +58,8 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
    */
   @Input() set widgetWidth(widgetWidth: number) {
     this._homePageOptions.widgetWidth = widgetWidth;
-    if (this.homePage) {
-      this.homePage.settings.widgetWidth = widgetWidth;
+    if (this.homepage) {
+      this.homepage.settings.widgetWidth = widgetWidth;
     }
   }
 
@@ -75,8 +72,8 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
    */
   @Input() set widgetHeight(widgetHeight: number) {
     this._homePageOptions.widgetHeight = widgetHeight;
-    if (this.homePage) {
-      this.homePage.settings.widgetHeight = widgetHeight;
+    if (this.homepage) {
+      this.homepage.settings.widgetHeight = widgetHeight;
     }
   }
 
@@ -89,8 +86,8 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
    */
   @Input() set animate(animate: boolean) {
     this._homePageOptions.animate = animate;
-    if (this.homePage) {
-      this.homePage.settings.animate = animate;
+    if (this.homepage) {
+      this.homepage.settings.animate = animate;
     }
   }
 
@@ -103,8 +100,8 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
    */
   @Input() set timeout(timeout: number) {
     this._homePageOptions.timeout = timeout;
-    if (this.homePage) {
-      this.homePage.settings.timeout = timeout;
+    if (this.homepage) {
+      this.homepage.settings.timeout = timeout;
     }
   }
 
@@ -117,8 +114,8 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
    */
   @Input() set easing(easing: EasingType) {
     this._homePageOptions.easing = easing;
-    if (this.homePage) {
-      this.homePage.settings.easing = easing;
+    if (this.homepage) {
+      this.homepage.settings.easing = easing;
     }
   }
 
@@ -132,7 +129,7 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
   private jQueryElement: JQuery;
 
   // Reference to the annotated SoHoXi control
-  private homePage: SohoHomePageStatic;
+  private homepage: SohoHomePageStatic;
 
   private _homePageOptions: SohoHomePageOptions = {};
 
@@ -140,6 +137,15 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
     private elementRef: ElementRef,
     private ngZone: NgZone,
   ) { }
+
+  /**
+   * Reset columns to their defaults (used on restore menu item).
+   */
+  refresh(animate?: boolean): void {
+    return this.ngZone.runOutsideAngular(() => {
+      this.homepage.refresh(animate);
+    });
+  }
 
   ngAfterViewInit() {
     this.ngZone.runOutsideAngular(() => {
@@ -153,7 +159,7 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
       // Once the control is initialised, extract the control
       // plug-in from the element.  The element name is
       // defined by the plug-in, but in this case is 'homepage'.
-      this.homePage = this.jQueryElement.data('homepage');
+      this.homepage = this.jQueryElement.data('homepage');
     });
   }
 
@@ -163,9 +169,9 @@ export class SohoHomePageComponent implements AfterViewInit, OnDestroy {
       if (this.jQueryElement) {
         this.jQueryElement.off();
       }
-      if (this.homePage) {
-        this.homePage.destroy();
-        this.homePage = null;
+      if (this.homepage) {
+        this.homepage.destroy();
+        this.homepage = null;
       }
     });
   }